+++ Job-Hunt/app/Http/Controllers/Admin/CompanyLocationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CompanyLocation;
+use App\Models\Company;
 
 class CompanyLocationController extends Controller
 {
@@ -58,6 +59,11 @@ class CompanyLocationController extends Controller
 
     public function delete($id)
     {
+          $check = Company::where('company_location_id',$id)->count();
+          if($check>0) 
+          {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+          }
           $company_location = CompanyLocation::find($id);
           $company_location ->delete();
