+++ Job-Hunt/app/Http/Controllers/Admin/JobLocationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Location;
+use App\Models\Job;
 
 class JobLocationController extends Controller
 {
@@ -58,6 +59,11 @@ class JobLocationController extends Controller
 
     public function delete($id)
     {
+          $check = Job::where('job_location_id',$id)->count();
+          if($check>0) 
+          {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+          }
           $job_location = Location::find($id);
           $job_location ->delete();
